+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Chore: move telegram link into toplevel of settings so that it can be found easier
 - Feat: update coordinator API to show more details on pending channel balance
+- Feat: show dlc-channel balance instead of ln-balance in app and in coordinator's API
 
 ## [1.7.4] - 2023-12-20